API/requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
fastapi==0.73.0
h11==0.13.0
httptools==0.3.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
mysql-connector-python==8.0.28
pkg_resources==0.0.0
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
